+++ releases.md
@@ -33,6 +33,7 @@
 - added `padLeft`, `padLeftWith`, `padRight` and `padRightWith` to the string module
 - added `.!` function to the core module
 - `novah run` now builds by default and requires `--no-build` otherwise
+- `novah build` now set up the base dependencies automatically
 
 ## Bug Fixes